small fix to autogenerate revision (--autogenerate)
