--- alembic/versions/30f75142008d_initial_revision.py
@@ -1,24 +1,0 @@
-"""initial revision
-
-Revision ID: 30f75142008d
-Revises: 
-Create Date: 2021-05-24 18:10:48.718930
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '30f75142008d'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    pass
-
-
-def downgrade():
-    pass
